# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Hexagonal blog platform REST API. Envelope responses use { ok, code, data, meta, error }.",
        "title": "Blog API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/turahe/blog-api"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/posts": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List admin posts",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "status filter",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "author UUID",
                        "name": "author_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "category UUID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create draft post",
                "parameters": [
                    {
                        "description": "draft post",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreatePost"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/posts/{param1}": {
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "post UUID",
                        "name": "param1",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "patch fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdatePost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/posts/{param1}/media": {
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Replace post media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "post UUID",
                        "name": "param1",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "media items",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.ReplacePostMedia"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/posts/{param1}/publish": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Publish post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "post UUID",
                        "name": "param1",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.CreatePost": {
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string",
                    "maxLength": 2000
                },
                "slug": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
        "requests.PostMediaItem": {
            "type": "object",
            "required": [
                "kind",
                "media_asset_id"
            ],
            "properties": {
                "kind": {
                    "type": "string"
                },
                "media_asset_id": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "requests.ReplacePostMedia": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "enforce_cover_consistency": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requests.PostMediaItem"
                    }
                }
            }
        },
...
        "requests.UpdatePost": {
            "type": "object",
            "properties": {
                "category_id": {
                    "description": "Category UUID; null clears the category, omitting the field keeps it.",
                    "type": "string",
                    "format": "uuid"
                },
                "content": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "responses.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "error": {
                    "$ref": "#/definitions/responses.ErrorBody"
                },
                "links": {
                    "$ref": "#/definitions/responses.PageLinks"
                },
                "meta": {
                    "type": "object"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "responses.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.PageLinks": {
            "type": "object",
            "properties": {
                "first": {
                    "type": "string"
                },
                "last": {
                    "type": "string"
                },
                "next": {
                    "type": "string"
                },
                "prev": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}